Fix integer size-related compiler warnings

+++ src/thermo/PlasmaPhase.cpp
@@ -68,7 +68,7 @@ void PlasmaPhase::updateElectronEnergyDistribution()
                 "Call to calculateDistributionFunction failed.");
         }
         bool validEEDF = (
-            m_electronEnergyDist.size() == m_nPoints &&
+            static_cast<size_t>(m_electronEnergyDist.size()) == m_nPoints &&
             m_electronEnergyDist.allFinite() &&
             m_electronEnergyDist.maxCoeff() > 0.0 &&
             m_electronEnergyDist.sum() > 0.0
